+++ docs/changes.rst
@@ -65,6 +65,7 @@ To be released.
   method.  [:issue:`98` by Lionel Koenig]
 - Started supporting resizing and cropping of GIF images.
   [:issue:`88` by Bear Dong]
+- Added :meth:`Image.flip() <wand.image.Image.flip>` method.
 
 __ http://en.wikipedia.org/wiki/Seam_carving